all error is < 0.07

+++ flash_attn/flash_attn_triton_amd/fwd_prefill.py
@@ -188,12 +188,12 @@ def _attn_fwd_inner(acc, l_i, m_i, q, k_ptrs, v_ptrs, bias_ptrs, stride_kn, stri
         acc = acc * alpha[:, None]
         if not PRE_LOAD_V:
             v = load_fn(v_ptrs, k_offs_n, k_offs_k, actual_seqlen_k, ACTUAL_BLOCK_DMODEL)
-            v = v.to(tl.float16)
+            # v = v.to(tl.float16)
         # -- update m_i and l_i
         l_i = l_i * alpha + l_ij
         # update m_i and l_i
         m_i = m_ij
-        acc += tl.dot(p.to(v.type.element_ty), v).to(tl.float16) #(v.to(tl.float32) * V_SCALE)
+        acc += tl.dot(p.to(v.type.element_ty), v) #(v.to(tl.float32) * V_SCALE)
         k_ptrs += BLOCK_N * stride_kn
         v_ptrs += BLOCK_N * stride_vk
         if bias_ptrs is not None:
